:truck: Move key codes table to common
